imp: add alternative display mode
This commit adds an alternative display mode for treemap. In the
original mode, leaf nodes are displayed inside the rectangle of their
parent, that is itself inside the rectangle of its parent, and so on.
With the new mode "headerBoxes" the parent rectangles do not contain
their children, but are displayed as headers above them.

The new mode is activated by setting the displayMode option to
"headerBoxes" in the dataset.